Structural finalization for the boyos + bootstrap logic
Extended the standard library into its final form, minus the IMU part (which shouldn't be that hard to go off), plus the bootstrap and a probably working game DL mechanism for the browser. This needs some tweaks to actually work, but the standard library seems to be raring to go.