Added most bootstrap code, tweaks to FPSR for exit strings
None of this is tested yet, but it should work nicely with few changes. EXIT pushes one argument to the stack and returns it as is; we're just swapping that to a character or zero and use that to arbitrate whether to actually sideload a game or not. Theoretically, this supports game-between-game calling - but it shouldn't be really ever used besides the browser.

Browser code is ready sans execution, waiting on Jessie to finish.