Pre-test commit of the game browser.
Not sure about some of the heap semantics here. We dynamically
allocate a 500-some byte array in the bootstrap, then wait
to free it until the call returns. TBA whether this will crash.

This also serves as the first big test of the string library +
the compiler in general beyond cherry-picked test cases.