Polished C library, tweaked the compiler to handle lines correctly, fixed nasty HTTPget bug
So the C library is squared away, for the most part - a few easy functions have yet to be implemented but they follow the same patterns as the ones done so far. It's safe to call this done, considering all classes of functions get run at least once during the game browser's execution.

About the game browser, we've reframed the concept of a 'host' on the server to encapsulate one person, including their Arduino. This means when you switch on the ESP32, you're greeted with a "henlo, would you like to run your own hosted game or browse for others?" Selecting the former downloads the game at your own host, the latter pops open the game browser downloaded from a hard-coded host that nobody in particular owns.

HTTP get was broken because it didn't spin for excessively long responses sent in a chunked fashion. This was resolved, after some more wrestling with client.available() and client.connected(). The result is a nice reference implementation that black boxes work away from the game devs!

Should be ready to mainline the games now. I'm investigating them now and they appear to be short enough to work. Going to switch them over to FPSR functions next week and get this done for good.