New and improved browser!
DId a number of things tonight. Utilized the finalized server API, added a hostname call to the FPSR library, and tightened up the bootloader.

Most importantly, the game browser has been completely revamped to support correct download/install semantics. Might be wise to add a primitive reference counter in the future, and will def need to tweak the supported source code size capacity. But it's alive!