Documentation + web app integration
Lots of new documentation - finalized the server/ readme and added the web app that allows for easier on-the-fly game posting. Could augment this with GETting + revisions later...

The README on the front page has a suitable discussion of the compiler, and the same is true of the server/ README re. the database API. Games is ready for integraton.