Retooled the hosting to fit server-side semantics
The hosting of old was exclusively per-device - meaning that a download of the game would copy-paste the code onto your host. This worked, but was prohibitive for multiplayer gaming. A layer of indirection was added on the server-side in host_handler.py, which allows temporary re-aliasing of hostnames and brings full compliance with server semantics.

Additionally, gethostname() was retooled to produce accurate results, and  updatehostname() was created. Due to security implementations, we really need to hide this call in the game runtime, but that's on a commit to come.