Added the billiards game, added random
Billiards works! Much suffering later, the game plays. Many style revisions are in order; we'll see those later tonight. Also expect another game, as indicated in the readme...

Randomint was implemented shadowing the random function. It's seeded in the bootstrap for you, should make sure that's extra clear in documentation to come.