Handful of bug fixes and deprecation of Post-It
The .NET API we used as a proof of concept web app seems to be yeeting
post data into the URI, and the documentation is poor enough that it's
worth deprecating in favor of a simple python test. The decision is much
more maintainable and also better for production usage, where you're
probably pulling from games in the existing tree rather than rolling
your own from scratch. I'll also be accepting PRs for new games far into
the future, so...

Also, there was a bug in the game browser that incorrectly allocated the
gameslist - resulting in potential buffer overflows given we add even a
few more hosts. Problem solved.

Bootstrap and libfpsr.h have been updated with better documentation and
are ready for week 4 game integration. Style is enforced project wide!